refactor: move vis config panels to shared folder
