+++ packages/frontend/src/components/VisualizationConfigs/ChartConfigPanel/ChartConfigPanel.styles.ts
