+++ packages/frontend/src/components/VisualizationConfigs/ChartConfigPanel/Legend/Legend.styles.tsx
