+++ packages/frontend/src/components/VisualizationConfigs/ChartConfigPanel/Legend/ReferenceLine.styles.tsx
