+++ packages/frontend/src/components/VisualizationConfigs/ChartConfigPanel/Series/InvalidSeriesConfiguration.tsx
