+++ packages/frontend/src/components/VisualizationConfigs/ChartConfigPanel/Series/Series.styles.ts
