+++ packages/frontend/src/components/VisualizationConfigs/TableConfigPanel/ConditionalFormatting.styles.ts
