+++ packages/frontend/src/components/VisualizationConfigs/TableConfigPanel/DroppableItemsList.tsx
