+++ packages/frontend/src/components/VisualizationConfigs/TableConfigPanel/GeneralSettings.tsx
@@ -3,8 +3,8 @@ import { Tooltip2 } from '@blueprintjs/popover2';
 import { Title } from '@mantine/core';
 import React, { FC, useCallback, useMemo, useState } from 'react';
 import { DragDropContext, DropResult } from 'react-beautiful-dnd';
-import useToaster from '../../hooks/toaster/useToaster';
-import { useVisualizationContext } from '../LightdashVisualization/VisualizationProvider';
+import useToaster from '../../../hooks/toaster/useToaster';
+import { useVisualizationContext } from '../../LightdashVisualization/VisualizationProvider';
 import ColumnConfiguration from './ColumnConfiguration';
 import DroppableItemsList from './DroppableItemsList';
