+++ packages/frontend/src/components/VisualizationConfigs/TableConfigPanel/TableConfig.styles.ts
