+++ packages/frontend/src/components/VisualizationConfigs/TableConfigPanel/index.tsx
@@ -3,9 +3,9 @@ import { Popover2 } from '@blueprintjs/popover2';
 import { Box, Button } from '@mantine/core';
 import { IconChevronDown } from '@tabler/icons-react';
 import React from 'react';
-import { COLLAPSABLE_CARD_BUTTON_PROPS } from '../common/CollapsableCard';
-import MantineIcon from '../common/MantineIcon';
-import { useVisualizationContext } from '../LightdashVisualization/VisualizationProvider';
+import { COLLAPSABLE_CARD_BUTTON_PROPS } from '../../common/CollapsableCard';
+import MantineIcon from '../../common/MantineIcon';
+import { useVisualizationContext } from '../../LightdashVisualization/VisualizationProvider';
 import ConditionalFormattingList from './ConditionalFormattingList';
 import GeneralSettings from './GeneralSettings';
